resolved the rest of the styled component layering
Moved the remaining styled component definitions outside of the component definition.

Beginning to extract the Add transaction for from the speed dial button component
